+++ src/compile/ruby/map_formulae_to_ruby.rb
@@ -29,6 +29,7 @@ class MapFormulaeToRuby < MapValuesToRuby
     :'COUNT' => 'count',
     :'COUNTA' => 'counta',
     :'COUNTIF' => 'countif',
+    :'COUNTIFS' => 'countifs',
     :'ENSURE_IS_NUMBER' => 'ensure_is_number',
     :'EXP' => 'exp',
     :'FIND' => 'find',