+++ src/excel/excel_functions.rb
@@ -140,3 +140,5 @@ require_relative 'excel_functions/floor'
 require_relative 'excel_functions/sqrt'
 
 require_relative 'excel_functions/product'
+
+require_relative 'excel_functions/countifs'